+++ src/main/java/com/almostreliable/summoningrituals/network/ClientAltarUpdatePacket.java
@@ -3,13 +3,13 @@ package com.almostreliable.summoningrituals.network;
 import com.almostreliable.summoningrituals.altar.AltarBlockEntity;
 import com.almostreliable.summoningrituals.util.Utils;
 import io.netty.buffer.Unpooled;
-import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.fabricmc.fabric.api.networking.v1.PacketSender;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientPacketListener;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.Level;
 
 public final class ClientAltarUpdatePacket {
 
@@ -42,12 +42,23 @@ public final class ClientAltarUpdatePacket {
         return buffer;
     }
 
-    public static void progressUpdate(BlockPos pos, int progress) {
-        ClientPlayNetworking.send(CHANNEL, encode(pos, progress, PacketType.PROGRESS));
+    public static void progressUpdate(
+        Level level, BlockPos pos, int progress,
+        Consumer4<Level, BlockPos, ResourceLocation, FriendlyByteBuf> packetSender
+    ) {
+        packetSender.accept(level, pos, CHANNEL, encode(pos, progress, PacketType.PROGRESS));
     }
 
-    public static void processTimeUpdate(BlockPos pos, int processTime) {
-        ClientPlayNetworking.send(CHANNEL, encode(pos, processTime, PacketType.PROCESS_TIME));
+    public static void processTimeUpdate(
+        Level level, BlockPos pos, int processTime,
+        Consumer4<Level, BlockPos, ResourceLocation, FriendlyByteBuf> packetSender
+    ) {
+        packetSender.accept(level, pos, CHANNEL, encode(pos, processTime, PacketType.PROCESS_TIME));
+    }
+
+    @SuppressWarnings("NewClassNamingConvention")
+    public interface Consumer4<T1, T2, T3, T4> {
+        void accept(T1 t1, T2 t2, T3 t3, T4 t4);
     }
 
     public enum PacketType {