+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## Unreleased
+- fixed dedicated server networking crash
+
 ## [2.0.0] - 2022-12-20
 
 Initial release of the Fabric port!